Csharp-Advanced/OOP/Exam Preparation/4.C# OOP Retake Exam - 18 April 2022/01. Structure_Skeleton/Heroes/Repositories/HeroRepository.cs
namespace Heroes.Repositories
{
    using System.Linq;
    using System.Collections.Generic;

    using Contracts;
    using Models.Contracts;

    public class HeroRepository : IRepository<IHero>
    {
        private List<IHero> heroes;
        public HeroRepository()
        {
            heroes = new List<IHero>();
        }
        public IReadOnlyCollection<IHero> Models
            => heroes.AsReadOnly();
        public void Add(IHero model)
            => heroes.Add(model);
        public IHero FindByName(string name)
            => heroes.FirstOrDefault(x => x.Name == name);
        public bool Remove(IHero model)
            => heroes.Remove(model);
    }
}
